requirements.txt
django>=3.0.2
psycopg2-binary
django-bootstrap4
ipython
django-filter